+++ adapters/Facebook/FacebookAdapter/GADMAdapterFacebook.m
@@ -57,8 +57,8 @@
 }
 
 - (instancetype)initWithGADMAdNetworkConnector:(id<GADMAdNetworkConnector>)connector {
-  if (![[GADMobileAds sharedInstance] isSDKVersionAtLeastMajor:7 minor:42 patch:2]) {
-    NSLog(@"Please use the latest version of the GoogleMobileAds SDK for best results.");
+  if (![[GADMobileAds sharedInstance] isSDKVersionAtLeastMajor:7 minor:46 patch:0]) {
+    NSLog(@"This version of the Facebook adapter requires a newer version of the Google Mobile Ads SDK.");
     return nil;
   }
   self = [self init];